% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfdr_function.R
\name{bfdr}
\alias{bfdr}
\title{Bayesian FDR function, based on ShrinkBayes::BFDR.}
\usage{
bfdr(x)
}
\arguments{
\item{x}{Posterior probabilities}
}
\value{
Corresponding Bayesian FDRs
}
\description{
Bayesian FDR function, based on ShrinkBayes::BFDR.
}
