% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{disp.prior.scale.log.posterior}
\alias{disp.prior.scale.log.posterior}
\title{Conditional log posterior function for scale parameter of dispersion prior. Called
by MCMC functions.}
\usage{
disp.prior.scale.log.posterior(
  prior.scale,
  prior.location,
  g,
  disps0,
  disps1,
  disps2,
  z
)
}
\arguments{
\item{prior.scale}{Current estimate of prior scale parameter}

\item{prior.location}{Current estimate of prior location parameter}

\item{g}{Number of genes}

\item{disps0}{Vector of current overall per-gene dispersion estimates}

\item{disps1}{Vector of current group 1 per-gene dispersion estimates}

\item{disps2}{Vector of current group 2 per-gene dispersion estimates}

\item{z}{Vector of current per-gene mixture component indicator variables}
}
\value{
Conditional posterior density of scale parameter of mean dispersion, on log
scale
}
\description{
Conditional log posterior function for scale parameter of dispersion prior. Called
by MCMC functions.
}
