% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_tail_prob_function.R
\name{hdi}
\alias{hdi}
\title{Highest density interval function, used in \code{hpd.pval()}.}
\usage{
hdi(x, credMass)
}
\arguments{
\item{x}{(Posterior) sample}

\item{credMass}{Desired density within interval}
}
\value{
(lower, upper) boundaries of the narrowest interval containing
the desired density
}
\description{
Based on \code{HDInterval::hdi()} package, but with sorting step removed since
sorting is done in \code{hpd.pval()}.
}
