% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_tail_prob_function.R
\name{hpd.pval}
\alias{hpd.pval}
\title{Function to calculate posterior tail probability that \code{x > m} using highest
density intervals.
If \code{x} is a posterior sample of the log-transformed difference in a given
parameter between groups, \code{hpd.pval()} estimates the posterior probability that
the log fold change in the parameter between groups is greater than the threshold
\code{m}.}
\usage{
hpd.pval(x, m = 0)
}
\arguments{
\item{x}{Posterior sample}

\item{m}{Threshold probability. Default is 0.}
}
\value{
Posterior probability that \code{x > m}
}
\description{
Function to calculate posterior tail probability that \code{x > m} using highest
density intervals.
If \code{x} is a posterior sample of the log-transformed difference in a given
parameter between groups, \code{hpd.pval()} estimates the posterior probability that
the log fold change in the parameter between groups is greater than the threshold
\code{m}.
}
