% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_hmm_one_chain_function.R
\name{ln_hmm_1_chain}
\alias{ln_hmm_1_chain}
\title{Function to run a single MCMC chain for the hierarchical mixture model.}
\usage{
ln_hmm_1_chain(
  counts,
  groups,
  chain.length,
  thin = 1,
  inits,
  mean.proposal.scales0 = rep(0.2, ncol(counts)),
  mean.proposal.scales1 = rep(0.2, ncol(counts)),
  mean.proposal.scales2 = rep(0.2, ncol(counts)),
  disp.proposal.scales0 = rep(0.5, ncol(counts)),
  disp.proposal.scales1 = rep(0.5, ncol(counts)),
  disp.proposal.scales2 = rep(0.5, ncol(counts)),
  mean.prior.scale.proposal.sd = 0.1,
  disp.prior.scale.proposal.sd = 0.4
)
}
\arguments{
\item{counts}{Matrix of counts with samples in rows and genes in columns. This will
be fixed in future versions to follow the convention of genes in rows and samples
in columns.}

\item{groups}{Factor indication group membership (1 or 2)}

\item{chain.length}{Number of iterations to run}

\item{thin}{Thinning. Default is 1, i.e. no thinning.}

\item{inits}{List of initial values for each parameter: overall and per-group means
(means0, means1, means2); overall and per-group dispersions (disps0, disps1,
disps2); mean hyperparameters (mean.prior.location, mean.prior.scale); dispersion
hyperparameters (disp.prior.location, disp.prior.scale).}

\item{mean.proposal.scales0}{Vector of scale parameters for overall mean proposal
densities. Default is 0.2.}

\item{mean.proposal.scales1}{Vector of scale parameters for group 1 mean proposal
densities Default is 0.2.}

\item{mean.proposal.scales2}{Vector of scale parameters for group 2 mean proposal
densities Default is 0.2.}

\item{disp.proposal.scales0}{Vector of scale parameters for overall dispersion
proposal densities Default is 0.5.}

\item{disp.proposal.scales1}{Vector of scale parameters for group 1 dispersion
proposal densities Default is 0.5.}

\item{disp.proposal.scales2}{Vector of scale parameters for group 2 dispersion
proposal densities Default is 0.5.}

\item{mean.prior.scale.proposal.sd}{Standard deviation for proposal density for mean
prior scale parameter. Default is 0.1.}

\item{disp.prior.scale.proposal.sd}{Standard deviation for proposal density for
dispersion prior scale parameter. Default is 0.4.}
}
\value{
List of inputs, acceptance rates and posterior samples. Inputs:
chain.length, thin, inits, mean.proposal.scales0, mean.proposal.scales1,
mean.proposal.scales2, disp.proposal.scales0, disp.proposal.scales1,
disp.proposal.scales2, mean.prior.scale.proposal.sd, disp.prior.scale.proposal.sd.
Overall and per-group acceptance rates for means and dispersions: accept.means0,
accept.means1, accept.means2, accept.disps0, accept.disps1, accept.disps2.
Acceptance rates for scale hyperparameters: accept.mean.prior.scale,
accept.disp.prior.scale. Matrices of overall and per-group posterior samples of
means and dispersions, with genes in columns: posterior.means0, posterior.means1,
posterior.means2, posterior.disps0, posterior.disps1, posterior.disps2. Vectors of
posterior samples of hyperparameters: posterior.mean.prior.location,
posterior.mean.prior.scale, posterior.disp.prior.location,
posterior.disp.prior.scale, posterior.indicators. Matrix of posterior samples of
mixture component indicators, with genes in columns: posterior.indicators. Vector
of posterior sample of proportion of differentially distributed genes:
posterior.proportion.
}
\description{
Called by \code{ln_hmm_adapt_3_chains}.
}
