% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_hmm_three_chains_function.R
\name{ln_hmm_3_chains}
\alias{ln_hmm_3_chains}
\title{Function to run three parallel MCMC chains for the hierarchical mixture model.}
\usage{
ln_hmm_3_chains(
  counts,
  groups,
  chain.length,
  thin = 1,
  inits1,
  inits2,
  inits3,
  mean.proposal.scales0 = rep(0.2, nrow(counts)),
  mean.proposal.scales1 = rep(0.2, nrow(counts)),
  mean.proposal.scales2 = rep(0.2, nrow(counts)),
  disp.proposal.scales0 = rep(0.5, nrow(counts)),
  disp.proposal.scales1 = rep(0.5, nrow(counts)),
  disp.proposal.scales2 = rep(0.5, nrow(counts)),
  mean.prior.scale.proposal.sd = 0.1,
  disp.prior.scale.proposal.sd = 0.4
)
}
\arguments{
\item{counts}{Matrix of counts with genes in rows and samples in columns}

\item{groups}{Factor indication group membership (1 or 2)}

\item{chain.length}{Number of iterations to run}

\item{thin}{Thinning. Default is 1, i.e. no thinning.}

\item{inits1}{List of initial values for each parameter for first chain:
overall and per-group means (means0, means1, means2); overall and per-group
dispersions (disps0, disps1, disps2); mean hyperparameters (mean.prior.location,
mean.prior.scale); dispersion hyperparameters (disp.prior.location,
disp.prior.scale).}

\item{inits2}{List of initial values for second chain.}

\item{inits3}{List of initial values for third chain.}

\item{mean.proposal.scales0}{Vector of scale parameters for overall mean proposal
densities. Default is 0.2.}

\item{mean.proposal.scales1}{Vector of scale parameters for group 1 mean proposal
densities Default is 0.2.}

\item{mean.proposal.scales2}{Vector of scale parameters for group 2 mean proposal
densities Default is 0.2.}

\item{disp.proposal.scales0}{Vector of scale parameters for overall dispersion
proposal densities Default is 0.5.}

\item{disp.proposal.scales1}{Vector of scale parameters for group 1 dispersion
proposal densities Default is 0.5.}

\item{disp.proposal.scales2}{Vector of scale parameters for group 2 dispersion
proposal densities Default is 0.5.}

\item{mean.prior.scale.proposal.sd}{Standard deviation for proposal density for mean
prior scale parameter. Default is 0.1.}

\item{disp.prior.scale.proposal.sd}{Standard deviation for proposal density for
dispersion prior scale parameter. Default is 0.4.}
}
\value{
List of inputs, acceptance rates and posterior samples. Inputs:
chain.length, thin, inits1, inits2, inits3, mean.proposal.scales0,
mean.proposal.scales1, mean.proposal.scales2, disp.proposal.scales0,
disp.proposal.scales1, disp.proposal.scales2, mean.prior.scale.proposal.sd,
disp.prior.scale.proposal.sd.
Overall and per-group acceptance rates for means and dispersions for each chain
(x = 1,2,3): accept.means0.x, accept.means1.x, accept.means2.x, accept.disps0.x,
accept.disps1.x, accept.disps2.x. Acceptance rates for scale hyperparameters:
accept.mean.prior.scale.x, accept.disp.prior.scale.x.
Matrices of overall and per-group posterior samples of
means and dispersions, with genes in columns: posterior.means0.x,
posterior.means1.x, posterior.means2.x, posterior.disps0.x, posterior.disps1.x,
posterior.disps2.x. Vectors of posterior samples of hyperparameters:
posterior.mean.prior.location.x, posterior.mean.prior.scale.x,
posterior.disp.prior.location.x, posterior.disp.prior.scale.x,
posterior.indicators.x. Matrices of posterior samples of mixture component
indicators, with genes in columns: posterior.indicators.x. Vectors of posterior
samples of proportion of differentially distributed genes: posterior.proportion.x.
}
\description{
Called by \code{ln_hmm_adapt_3_chains}.
}
