% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/lognormal_hmm_adaptive_proposals_three_chains_function.R
\name{ln_hmm_adapt_3_chains}
\alias{ln_hmm_adapt_3_chains}
\title{Function to run overall MCMC algorithm.}
\usage{
ln_hmm_adapt_3_chains(
  counts,
  groups,
  chain.length = 2000,
  initial.chain.length = 100,
  adapt.chain.length = 100
)
}
\arguments{
\item{counts}{Matrix of counts with samples in rows and genes in columns. This will
be fixed in future versions to follow the convention of genes in rows and samples
in columns.}

\item{groups}{Factor indication group membership (1 or 2)}

\item{chain.length}{Number of iterations to run in final phase. Default is 2000.}

\item{initial.chain.length}{Number of iterations to run in initial phase. Default is
100.}

\item{adapt.chain.length}{Number of MCMC iterations to run in each iteration of
adaptive phase. Default is 100.}
}
\value{
List of supplied parameters and results. Length of final MCMC runs
(chain.length). Number of MCMC runs taken in adaptive phase (adaptive.runs).
Scale/standard deviation parameters for proposal densities in final run
(mean.proposal.scales0, mean.proposal.scales1, mean.proposal.scales2,
disp.proposal.scales0, disp.proposal.scales1, disp.proposal.scales2,
mean.prior.scale.proposal.sd, disp.prior.scale.proposal.sd). \code{mcmc.list} objects
containing \code{mcmc} objects for posterior samples from each chain for each
parameter (means0, means1,means2, disps0, disps1, disps2, mean.prior.location,
disp.prior.location,mean.prior.scale, disp.prior.scale, indicators, proportion).
}
\description{
Initial phase: Short single chain to get rough estimates of range of values for each
parameter, used to obtain overdispersed initial values for adaptive phase.
Adaptive phase: Three parallel chains, initially using overdispersed initial values
from initial phase, run iteratively to obtain proposal densities that give acceptance
rates within the desired range. Parameter values continue through iterations, i.e.
final samples for each iteration are used as initial values for the next iteration.
Final phase: Three parallel chains, used to obtain posterior samples for each
parameter. Proposal densities taken from adaptive phase, and initial parameter values
are final values for each chain from the adaptive phase.
Posterior samples are returned as \code{mcmc.list} objects containing \code{mcmc}
objects from the \code{coda} package for each final phase MCMC chain.
}
