% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{mean.conditional.log.posterior}
\alias{mean.conditional.log.posterior}
\title{Per-gene conditional log posterior mean function.}
\usage{
\method{mean}{conditional.log.posterior}(n, sample.means, means, disps, prior.location, prior.scale)
}
\arguments{
\item{n}{Number of samples}

\item{sample.means}{Vector of per-gene sample means}

\item{means}{Vector of current per-gene mean estimates}

\item{disps}{Vector of current per-gene dispersion estimates}

\item{prior.location}{Current estimate of prior location parameter}

\item{prior.scale}{Current estimate of prior scale parameter}
}
\value{
Vector of conditional posterior mean densities on log scale
}
\description{
Called by MCMC functions.
}
