% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{prior.location.posterior.mean}
\alias{prior.location.posterior.mean}
\title{Function to compute mean of conditional posteriors for prior location parameters.
Used for Gibbs sampling for prior location parameters.}
\usage{
prior.location.posterior.mean(
  hyperprior.mean,
  hyperprior.var,
  prior.scale,
  parameters0,
  parameters1,
  parameters2,
  z
)
}
\arguments{
\item{hyperprior.mean}{Mean of hyperprior on prior location parameter}

\item{hyperprior.var}{Variance of hyperprior on prior location parameter}

\item{prior.scale}{Current estimate of prior scale parameter}

\item{parameters0}{Vector of current overall mean or dispersion estimates}

\item{parameters1}{Vector of current group 1 mean or dispersion estimates}

\item{parameters2}{Vector of current group 2 mean or dispersion estimates}

\item{z}{Vector of current per-gene mixture component indicator variables}
}
\value{
Mean of conditional posterior distribution of the given prior location
parameter
}
\description{
Function to compute mean of conditional posteriors for prior location parameters.
Used for Gibbs sampling for prior location parameters.
}
