% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{prior.location.posterior.sd}
\alias{prior.location.posterior.sd}
\title{Function to compute standard deviation of conditional posteriors for prior
location parameters. Used for Gibbs sampling for prior location parameters.}
\usage{
prior.location.posterior.sd(hyperprior.var, prior.scale, z)
}
\arguments{
\item{hyperprior.var}{Variance of hyperprior on prior location parameter}

\item{prior.scale}{Current estimate of prior scale parameter}

\item{z}{Vector of current per-gene mixture component indicator variables}
}
\value{
Standard deviation of conditional posterior distribution of the given prior
location parameter
}
\description{
Function to compute standard deviation of conditional posteriors for prior
location parameters. Used for Gibbs sampling for prior location parameters.
}
