% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{pz0}
\alias{pz0}
\title{Function for raw conditional log posterior probability of no differential
distribution. Called by \code{posterior.indicator.probabilities()}.}
\usage{
pz0(
  genes,
  counts,
  n,
  sample.means,
  means,
  disps,
  mean.prior.location,
  mean.prior.scale,
  disp.prior.location,
  disp.prior.scale,
  lambda
)
}
\arguments{
\item{genes}{Number of genes}

\item{counts}{Matrix of counts with genes in rows and samples in columns}

\item{n}{Number of samples}

\item{sample.means}{Vector of per-gene sample means}

\item{means}{Vector of current per-gene mean estimates}

\item{disps}{Vector of current per-gene dispersion estimates}

\item{mean.prior.location}{Current estimate of mean prior location parameter}

\item{mean.prior.scale}{Current estimate of mean prior scale parameter}

\item{disp.prior.location}{Current estimate of dispersion prior location parameter}

\item{disp.prior.scale}{Current estimate of dispersion prior scale parameter}

\item{lambda}{Current estimate of proportion of differentially distributed genes}
}
\value{
Vector of un-normalised per-gene conditional posterior probabilities of no
differential distribution, on log scale
}
\description{
Function for raw conditional log posterior probability of no differential
distribution. Called by \code{posterior.indicator.probabilities()}.
}
