% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_posterior_functions_lognormal_hmm.R
\name{pz1}
\alias{pz1}
\title{Function for raw conditional log posterior probability of differential
distribution. Called by \code{posterior.indicator.probabilities()}.}
\usage{
pz1(
  genes,
  counts1,
  counts2,
  n1,
  n2,
  sample.means1,
  sample.means2,
  means1,
  means2,
  disps1,
  disps2,
  mean.prior.location,
  mean.prior.scale,
  disp.prior.location,
  disp.prior.scale,
  lambda
)
}
\arguments{
\item{genes}{Number of genes}

\item{counts1}{Matrix of counts in group 1 with genes in rows and samples in columns}

\item{counts2}{Matrix of counts in group 2}

\item{n1}{Number of samples in group 1}

\item{n2}{Number of samples in group 2}

\item{sample.means1}{Vector of per-gene group 1 sample means}

\item{sample.means2}{Vector of per-gene group 2 sample means}

\item{means1}{Vector of current per-gene group 1 mean estimates}

\item{means2}{Vector of current per-gene group 2 mean estimates}

\item{disps1}{Vector of current per-gene group 1 dispersion estimates}

\item{disps2}{Vector of current per-gene group 2 dispersion estimates}

\item{mean.prior.location}{Current estimate of mean prior location parameter}

\item{mean.prior.scale}{Current estimate of mean prior scale parameter}

\item{disp.prior.location}{Current estimate of dispersion prior location parameter}

\item{disp.prior.scale}{Current estimate of dispersion prior scale parameter}

\item{lambda}{Current estimate of proportion of differentially distributed genes}
}
\value{
Vector of un-normalised per-gene conditional posterior probabilities of
differential distribution, on log scale
}
\description{
Function for raw conditional log posterior probability of differential
distribution. Called by \code{posterior.indicator.probabilities()}.
}
